AFLGuidance: Report timeouts from any thread
If a timeout occurrs in a thread other than the thread that has
invoked the test method, then the TimeoutException does not
propagate to the FuzzStatement and can remain uncaught.

Therefore, this commit adds a flag that is set once per test
run whenever a timeout is detected in any thread. If the flag
is set at the end of the run, then the result is forcibly
changed to TIMEOUT.

This commit also bundles a test for verifying that the result is
a TIMEOUT when the exception is actually caught.

Addresses #26.