AFLGuidance: Allocate traceBits at init time
This change allocates the traceBits buffer used for storing coverage info when AFLGuidance instance is created,
and simply fills it up with zeros before every new test run, as opposed to allocating a new buffer for each
execution.

A subtle effect of this change is that it ensures that traceBits is never null, including before the first
test execution. This may be important if class loading activity fires off coverage events before the test
method is invoked, something that became important after commit 65a9f09.

This issue therefore should also fix #140, which was caused by NullPointerException being thrown
before the first test execution.